Put linkxxx variables into AnalyticsClick
